requirements.txt
openai-whisper==20240930
torch==2.4.1
pytest

